Update CryptoService.cs to newer Mono.Cecil version that's FIPS compliant
ILRepack uses an older version of Mono.Cecil that is not FIPS compliant.  The latest Mono.Cecil source is FIPS compliant, but ILRepack would need some semi-major changes to work with the new Mono.Cecil source so this is a band-aid until that can be done.  Note, this is NOT my code, I obtained it from https://github.com/KirillOsenkov/cecil/tree/dev/kirillo/reentrancy/Mono.Security.Cryptography